2-mygitrepos-summary/requirements.txt
smolagents>=1.4.1
requests>=2.32.3
python-dotenv>=1.0.1

